Cleaned code, added functionality and removed RDF prep function
Rewrote the code such that the entire workflow process, supertools, and simple tools are all considered actions. An action is a hub (a node with its edges and directly-related nodes) with extra attributes, including other actions it composes and is composed of.  Removed RDF functions, because these are deprecated after the changes