Reorganized the .graphml parsing code. Generated .ttl files
Added the Schema class.  An action may have multiple schemas and each schema holds a sequence of actions and artefacts by which the action can be defined.

The .graphml is a representation of the workflow's schema. An action representing the entire workflow can be added by providing a 'trapezoid' node with the workflow name in the .graphml and running add_context_edges(graph). Then, by instantiating Action(<wfname>, graph), the action is defined in terms of its sub-actions and artefacts. This is iterative, so if sub-actions have sub-actions, a new schema is generated for them as well.

Note: Currently, all nodes are part of each schema, which causes overlaps of actions (E.g., addField -> CalculateField and addAndCalculateField may be in the same schema). These overlaps can be spotted and potentially removed by matching inputs and outputs between schema edges, but the overlaps can also be considered a feature.