Removed numbering of unlabeled inputs
Adjusted the python script and the .ttl files such that if an edge between an input and a tool in a .graphml does not have a label, then it will receive the predicate 'inputx'. If there is a label, it will receive that label (There can be no duplicate labels per tool).